+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - prevent duplicate submissions to one upload
 - restrict number of submissions
 - time restrictions
+- improve received points
 
 
 # 0.2.7-SNAPSHOT